requirements.txt
fastapi
pyyaml==6.0.1
python-dateutil==2.9.0.post0
langdetect==1.0.9
uvicorn[standard]
pydantic[dotenv]
pydantic-settings
openai>=1.50.0
anthropic>=0.40.0 
tiktoken
python-dotenv==1.0.1
httpx
python-multipart
twilio
unidecode
psycopg[binary]
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
langchain-text-splitters>=0.2.0
pymupdf==1.24.5
Pillow==10.3.0
anthropic==0.28.0
langgraph==0.2.28
langchain-core==0.3.0
matplotlib
faker